refactor: move more items to core
Move more classes to the shared core module, instead of hanging around in the app module
