+++ app/src/main/java/com/chesire/malime/flow/login/syncing/SyncingFragment.kt
@@ -9,7 +9,7 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.get
 import androidx.navigation.fragment.findNavController
 import com.chesire.lifecyklelog.LogLifecykle
-import com.chesire.malime.AsyncState
+import com.chesire.malime.core.flags.AsyncState
 import com.chesire.malime.databinding.FragmentSyncingBinding
 import com.chesire.malime.extensions.hide
 import com.chesire.malime.extensions.show
@@ -48,9 +48,9 @@ class SyncingFragment : DaggerFragment() {
             viewLifecycleOwner,
             Observer {
                 when (it) {
-                    is AsyncState.Success -> findNavController().navigate(SyncingFragmentDirections.toAnimeFragment())
-                    is AsyncState.Loading -> setLoading()
-                    is AsyncState.Error -> hideLoading()
+                    is com.chesire.malime.core.flags.AsyncState.Success -> findNavController().navigate(SyncingFragmentDirections.toAnimeFragment())
+                    is com.chesire.malime.core.flags.AsyncState.Loading -> setLoading()
+                    is com.chesire.malime.core.flags.AsyncState.Error -> hideLoading()
                 }
             }
         )
