+++ app/src/main/java/com/chesire/malime/flow/series/search/SearchViewModel.kt
@@ -4,16 +4,16 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.chesire.malime.AsyncState
+import com.chesire.malime.core.flags.AsyncState
 import com.chesire.malime.AuthCaster
 import com.chesire.malime.server.Resource
 import com.chesire.malime.server.api.SearchApi
 import com.chesire.malime.core.flags.SeriesType
 import com.chesire.malime.core.flags.UserSeriesStatus
 import com.chesire.malime.core.models.SeriesModel
-import com.chesire.malime.extensions.postError
-import com.chesire.malime.extensions.postLoading
-import com.chesire.malime.extensions.postSuccess
+import com.chesire.malime.core.extensions.postError
+import com.chesire.malime.core.extensions.postLoading
+import com.chesire.malime.core.extensions.postSuccess
 import com.chesire.malime.repo.SeriesRepository
 import kotlinx.coroutines.launch
 import javax.inject.Inject
@@ -23,12 +23,12 @@ class SearchViewModel @Inject constructor(
     private val search: SearchApi,
     private val authCaster: AuthCaster
 ) : ViewModel() {
-    private val _searchResults = MutableLiveData<AsyncState<List<SeriesModel>, SearchError>>()
+    private val _searchResults = MutableLiveData<com.chesire.malime.core.flags.AsyncState<List<SeriesModel>, SearchError>>()
 
     val searchTitle = MutableLiveData<String>()
     val series: LiveData<List<SeriesModel>>
         get() = repo.series
-    val searchResults: LiveData<AsyncState<List<SeriesModel>, SearchError>>
+    val searchResults: LiveData<com.chesire.malime.core.flags.AsyncState<List<SeriesModel>, SearchError>>
         get() = _searchResults
 
     var seriesType: SeriesType = SeriesType.Anime
