+++ app/src/test/java/com/chesire/malime/flow/login/syncing/SyncingViewModelTests.kt
@@ -2,7 +2,7 @@ package com.chesire.malime.flow.login.syncing
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Observer
-import com.chesire.malime.AsyncState
+import com.chesire.malime.core.flags.AsyncState
 import com.chesire.malime.CoroutinesMainDispatcherRule
 import com.chesire.malime.server.Resource
 import com.chesire.malime.repo.SeriesRepository
@@ -30,14 +30,14 @@ class SyncingViewModelTests {
             coEvery { refreshAnime() } coAnswers { Resource.Error("") }
             coEvery { refreshManga() } coAnswers { Resource.Success(mockk()) }
         }
-        val mockObserver = mockk<Observer<AsyncState<Any, Any>>> {
+        val mockObserver = mockk<Observer<com.chesire.malime.core.flags.AsyncState<Any, Any>>> {
             every { onChanged(any()) } just Runs
         }
 
         SyncingViewModel(mockRepo).run {
             syncStatus.observeForever(mockObserver)
             syncLatestData()
-            assertTrue(syncStatus.value is AsyncState.Error)
+            assertTrue(syncStatus.value is com.chesire.malime.core.flags.AsyncState.Error)
         }
     }
 
@@ -47,14 +47,14 @@ class SyncingViewModelTests {
             coEvery { refreshAnime() } coAnswers { Resource.Success(mockk()) }
             coEvery { refreshManga() } coAnswers { Resource.Error("") }
         }
-        val mockObserver = mockk<Observer<AsyncState<Any, Any>>> {
+        val mockObserver = mockk<Observer<com.chesire.malime.core.flags.AsyncState<Any, Any>>> {
             every { onChanged(any()) } just Runs
         }
 
         SyncingViewModel(mockRepo).run {
             syncStatus.observeForever(mockObserver)
             syncLatestData()
-            assertTrue(syncStatus.value is AsyncState.Error)
+            assertTrue(syncStatus.value is com.chesire.malime.core.flags.AsyncState.Error)
         }
     }
 
@@ -64,14 +64,14 @@ class SyncingViewModelTests {
             coEvery { refreshAnime() } coAnswers { Resource.Success(mockk()) }
             coEvery { refreshManga() } coAnswers { Resource.Success(mockk()) }
         }
-        val mockObserver = mockk<Observer<AsyncState<Any, Any>>> {
+        val mockObserver = mockk<Observer<com.chesire.malime.core.flags.AsyncState<Any, Any>>> {
             every { onChanged(any()) } just Runs
         }
 
         SyncingViewModel(mockRepo).run {
             syncStatus.observeForever(mockObserver)
             syncLatestData()
-            assertTrue(syncStatus.value is AsyncState.Success)
+            assertTrue(syncStatus.value is com.chesire.malime.core.flags.AsyncState.Success)
         }
     }
 }
