+++ core/src/main/java/com/chesire/malime/core/extensions/MutableLiveDataExtensions.kt
@@ -1,25 +1,25 @@
-package com.chesire.malime.extensions
+package com.chesire.malime.core.extensions
 
 import androidx.lifecycle.MutableLiveData
-import com.chesire.malime.AsyncState
+import com.chesire.malime.core.flags.AsyncState
 
 /**
  * Posts to the receiver with [AsyncState.Success].
  */
 fun <T, E> MutableLiveData<AsyncState<T, E>>.postSuccess(data: T) {
-    postValue(AsyncState.Success(data))
+    postValue(com.chesire.malime.core.flags.AsyncState.Success(data))
 }
 
 /**
  * Posts to the receiver with [AsyncState.Error].
  */
 fun <T, E> MutableLiveData<AsyncState<T, E>>.postError(error: E) {
-    postValue(AsyncState.Error(error))
+    postValue(com.chesire.malime.core.flags.AsyncState.Error(error))
 }
 
 /**
  * Posts to the receiver with [AsyncState.Loading].
  */
 fun <T, E> MutableLiveData<AsyncState<T, E>>.postLoading() {
-    postValue(AsyncState.Loading())
+    postValue(com.chesire.malime.core.flags.AsyncState.Loading())
 }
